+++ src/components/cardOrder/cardOrder.module.css
@@ -1,0 +1,7 @@
+.input {
+  grid-area: input;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 0;
+}
