+++ src/components/menuProfile/menuProfile.module.css
@@ -1,0 +1,21 @@
+.nav {
+  grid-area: nav;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+}
+
+.link {
+  height: 64px;
+  text-decoration: none;
+  cursor: pointer;
+  transition: color 1s linear;
+}
+
+.link:hover {
+  color: #F2F2F3;
+}
+
+.link_activе {
+  color: #F2F2F3;
+}
