+++ src/pages/orders.module.css
@@ -1,9 +1,20 @@
 .orders {
+  display: grid;
+  grid-template-areas: "nav input item-1"
+                       "text item-2 item-3";
+  grid-template-rows: repeat(2, 1fr);
+  grid-template-columns: repeat(3, 1fr);
+  max-width: 1240px;
+  min-width: 480px;
+  background-color: #131316;
+  margin: 0 auto;
+  margin-top: 180px;
+}
+
+.cardOrder {
+  grid-area: input;
   display: flex;
   flex-direction: column;
   align-items: center;
-  max-width: 100vw;
-  min-width: 480px;
-  background-color: #131316;
-  margin-top: 180px;
+  padding: 0;
 }
