+++ src/pages/profile.module.css
@@ -11,30 +11,8 @@
   margin-top: 180px;
 }
 
-.nav {
-  grid-area: nav;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-}
-
 .form {
   max-width: 320px;
-}
-
-.link {
-  height: 64px;
-  text-decoration: none;
-  cursor: pointer;
-  transition: color 1s linear;
-}
-
-.link:hover {
-  color: #F2F2F3;
-}
-
-.link_activе {
-  color: #F2F2F3;
 }
 
 .input {
